project_root/requirements-security.txt
# Security-focused dependencies for the Enterprise Reporting System

# Core security libraries
cryptography>=3.4.8,<4.0.0
pycryptodome>=3.15.0,<4.0.0

# Password hashing and security
passlib>=1.7.4,<2.0.0
bcrypt>=3.2.0,<4.0.0

# JWT handling with security considerations
PyJWT>=2.4.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Input validation and sanitization
bleach>=5.0.0,<6.0.0
html-sanitizer>=1.0.0,<2.0.0

# Secure session management
itsdangerous>=2.1.0,<3.0.0

# Security headers for web applications
secure>=0.3.0,<1.0.0

# Rate limiting to prevent abuse
limits>=2.5.0,<3.0.0
flask-limiter>=2.6.0,<3.0.0

# Security-focused HTTP client
requests>=2.28.0,<3.0.0
urllib3>=1.26.0,<2.0.0

# Vulnerability scanning tools (development dependencies)
bandit>=1.7.0,<2.0.0
safety>=2.3.0,<3.0.0
semgrep>=1.0.0,<2.0.0
pip-audit>=0.4.0,<1.0.0

# Secure file handling
python-magic>=0.4.27,<1.0.0

# Certificate validation
certifi>=2022.0.0,<2023.0.0

# Secure random number generation
secrets>=1.0.0,<2.0.0